Implement constrained primitive types

+++ src/kio/schema/add_partitions_to_txn/response/v1.py
@@ -6,14 +6,16 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import TopicName
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AddPartitionsToTxnPartitionResult:
     __flexible__: ClassVar[bool] = False
-    partition_index: int = field(metadata={"kafka_type": "int32"})
+    partition_index: i32 = field(metadata={"kafka_type": "int32"})
     """The partition indexes."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The response error code."""
 
 
@@ -29,7 +31,7 @@ class AddPartitionsToTxnTopicResult:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class AddPartitionsToTxnResponse:
     __flexible__: ClassVar[bool] = False
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """Duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
     results: tuple[AddPartitionsToTxnTopicResult, ...]
     """The results for each topic."""
