+++ src/kio/schema/create_delegation_token/request/v2.py
@@ -5,6 +5,8 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i64
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class CreatableRenewers:
@@ -20,5 +22,5 @@ class CreateDelegationTokenRequest:
     __flexible__: ClassVar[bool] = True
     renewers: tuple[CreatableRenewers, ...]
     """A list of those who are allowed to renew this token before it expires."""
-    max_lifetime_ms: int = field(metadata={"kafka_type": "int64"})
+    max_lifetime_ms: i64 = field(metadata={"kafka_type": "int64"})
     """The maximum lifetime of the token in milliseconds, or -1 to use the server side default."""
