+++ src/kio/schema/create_topics/response/v0.py
@@ -6,6 +6,7 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import TopicName
+from kio.schema.primitive import i16
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -13,7 +14,7 @@ class CreatableTopicResult:
     __flexible__: ClassVar[bool] = False
     name: TopicName = field(metadata={"kafka_type": "string"})
     """The topic name."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The error code, or 0 if there was no error."""
