+++ src/kio/schema/create_topics/response/v6.py
@@ -6,6 +6,9 @@ from dataclasses import field
 from typing import ClassVar
 
 from kio.schema.entity import TopicName
+from kio.schema.primitive import i8
+from kio.schema.primitive import i16
+from kio.schema.primitive import i32
 
 
 @dataclass(frozen=True, slots=True, kw_only=True)
@@ -17,7 +20,7 @@ class CreatableTopicConfigs:
     """The configuration value."""
     read_only: bool = field(metadata={"kafka_type": "bool"})
     """True if the configuration is read-only."""
-    config_source: int = field(metadata={"kafka_type": "int8"}, default=-1)
+    config_source: i8 = field(metadata={"kafka_type": "int8"}, default=i8(-1))
     """The configuration source."""
     is_sensitive: bool = field(metadata={"kafka_type": "bool"})
     """True if this configuration is sensitive."""
@@ -28,15 +31,15 @@ class CreatableTopicResult:
     __flexible__: ClassVar[bool] = True
     name: TopicName = field(metadata={"kafka_type": "string"})
     """The topic name."""
-    error_code: int = field(metadata={"kafka_type": "int16"})
+    error_code: i16 = field(metadata={"kafka_type": "int16"})
     """The error code, or 0 if there was no error."""
     error_message: str | None = field(metadata={"kafka_type": "string"})
     """The error message, or null if there was no error."""
-    topic_config_error_code: int = field(metadata={"kafka_type": "int16"})
+    topic_config_error_code: i16 = field(metadata={"kafka_type": "int16"})
     """Optional topic config error returned if configs are not returned in the response."""
-    num_partitions: int = field(metadata={"kafka_type": "int32"}, default=-1)
+    num_partitions: i32 = field(metadata={"kafka_type": "int32"}, default=i32(-1))
     """Number of partitions of the topic."""
-    replication_factor: int = field(metadata={"kafka_type": "int16"}, default=-1)
+    replication_factor: i16 = field(metadata={"kafka_type": "int16"}, default=i16(-1))
     """Replication factor of the topic."""
     configs: tuple[CreatableTopicConfigs, ...]
     """Configuration of the topic."""
@@ -45,7 +48,7 @@ class CreatableTopicResult:
 @dataclass(frozen=True, slots=True, kw_only=True)
 class CreateTopicsResponse:
     __flexible__: ClassVar[bool] = True
-    throttle_time_ms: int = field(metadata={"kafka_type": "int32"})
+    throttle_time_ms: i32 = field(metadata={"kafka_type": "int32"})
     """The duration in milliseconds for which the request was throttled due to a quota violation, or zero if the request did not violate any quota."""
     topics: tuple[CreatableTopicResult, ...]
     """Results for each topic we tried to create."""
