+++ src/kio/schema/delete_acls/request/v1.py
@@ -5,23 +5,25 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import ClassVar
 
+from kio.schema.primitive import i8
+
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class DeleteAclsFilter:
     __flexible__: ClassVar[bool] = False
-    resource_type_filter: int = field(metadata={"kafka_type": "int8"})
+    resource_type_filter: i8 = field(metadata={"kafka_type": "int8"})
     """The resource type."""
     resource_name_filter: str | None = field(metadata={"kafka_type": "string"})
     """The resource name."""
-    pattern_type_filter: int = field(metadata={"kafka_type": "int8"}, default=3)
+    pattern_type_filter: i8 = field(metadata={"kafka_type": "int8"}, default=i8(3))
     """The pattern type."""
     principal_filter: str | None = field(metadata={"kafka_type": "string"})
     """The principal filter, or null to accept all principals."""
     host_filter: str | None = field(metadata={"kafka_type": "string"})
     """The host filter, or null to accept all hosts."""
-    operation: int = field(metadata={"kafka_type": "int8"})
+    operation: i8 = field(metadata={"kafka_type": "int8"})
     """The ACL operation."""
-    permission_type: int = field(metadata={"kafka_type": "int8"})
+    permission_type: i8 = field(metadata={"kafka_type": "int8"})
     """The permission type."""
